Upgrade all interfaces to use the new protocol
