+++ src/aio.rs
@@ -94,15 +94,12 @@ macro_rules! impl_async_methods {
                         }
                         Err(e) => match e {
                             ParseError::NotEnough => (),
-                            ParseError::BadPacket | ParseError::UnexpectedByte => {
+                            ParseError::BadPacket => {
                                 self.buffer.clear();
                                 return Err(SkyhashError::InvalidResponse.into());
                             }
                             ParseError::DataTypeError => {
                                 return Err(SkyhashError::ParseError.into())
-                            }
-                            ParseError::Empty => {
-                                return Err(IoError::from(ErrorKind::ConnectionReset).into())
                             }
                             ParseError::UnknownDatatype => {
                                 return Err(SkyhashError::UnknownDataType.into())
@@ -113,11 +110,7 @@ macro_rules! impl_async_methods {
             }
             /// This function is a subroutine of `run_query` used to parse the response packet
             fn try_response(&mut self) -> Result<(RawResponse, usize), ParseError> {
-                if self.buffer.is_empty() {
-                    // The connection was possibly reset
-                    return Err(ParseError::Empty);
-                }
-                Parser::new(&self.buffer).parse()
+                Parser::parse(&self.buffer)
             }
         }
         impl crate::actions::AsyncSocket for $ty {
