+++ src/lib.rs
@@ -229,7 +229,6 @@ pub mod ddl;
 pub mod error;
 pub mod pool;
 pub mod types;
-pub mod v2;
 // endof public mods
 // private mods
 mod deserializer;
